feat(#3) network scrypt
add condition for remove file create by cloud-init

+++ vsphere/scrypt/network.sh
@@ -11,5 +11,7 @@ network:
       nameservers:
         addresses: [{{DNS}}]
 EOF
-rm /etc/netplan/50-cloud-init.yaml
+if [ -f /etc/netplan/50-cloud-init.yaml ]; then
+  rm /etc/netplan/50-cloud-init.yaml
+fi
 netplan apply
